Update to go-libaudit v2.0.0 (#19300)
* Update to go-libaudit v2.0.0

This updates the go-libaudit library to v2.0.0. Some of the updates since the last release were already
incorporated into Beats.

https://github.com/elastic/go-libaudit/releases/tag/v2.0.0

Fixes #14055 (seccomp field enrichments)
Fixes a few minor typos and spelling issues in auditd event.category and event.action values.

* Only import go-libaudit on linux

(cherry picked from commit 4f3d9507f51f16476971aee16bd76fbaa40f195d)
